Initial header files added
Each of these header files calls the Anita header files and inherits classes from those to create a new class for the Binned analysis.  A BinnedResponseManager.h is not included in this initial upload.  The modification to ResponseManager.h is only a single additional include statement which is not used and should be moved to the .c file or the analysis code if it is needed.